Revised interface.
Changed interfaces to set data members at construction if they are common to
multiple methods or classes.  This led to shorter argument lists and less
redundancy for methods.  Common data members are set by the ModalDecomp
constructor.  Derived class constructors add additional data members.  A
private method for computing/decomposing Hankel matrices has been added so
that POD and BPOD can make use of them as needed.  (POD will do so with the
the same set of snapshots for direct and adjoint.)  This was done because
this implementation requires passing in the snapshot path lists as
arguments, but the public interface passes these in at construction only.