Added the low-level functions to bpod.py
I added a lot of code to BPOD so it is nearly complete. There are still many
bugs though. In particular, I didn't check yet that all of the variables match
with what Jon originally wrote in the skeleton. Some may be missing or have an extra
leading underscore, for example.