Corrected many of the small inconsistencies with the skeleton and low-level functions.
After discussing with Jon, we corrected some of the interfaces and public/private
data members. test.py runs now and fails where it should (because of empty matrices),
though there are still probably more errors to  correct. Also, more care needs to
be put into how to deal with the maximum memory size. Right now it's just a parameter
maxSnapsInMem but it would be nice if it was determined automatically.