Parallel tests pass for modaldeocmp
I have _compute_modes working in serial and parallel now. It is
thoroughly tested in testmodaldecomp.py. It could be done a little
more flexibly with parallel/serial cases, but as it is all one has to
do to test is:
python testmodaldecomp.py
mpiexec -n <# procs> python testmodaldecomp.py

It would be nice if only the mpiexec command was necessary in the future.