modaldecomp is now completely functional. the util.MPI class will not fail without mpi4py
the modaldecomp is completely tested in serial and parallel. It will work completely
without mpi4py installed without errors as long as it is run in serial. It will also
efficiently work in parallel if mpi4py is installed and used with mpiexec. The MPI
class has a few changes also to make it work when mpi4py is unavailable.

The biggest change in this update is that _compute_inner_product_chunk is now tested and
passes all tests. It should be possible to write all of BPOD, POD, and DMD now.