Added tests for DMD class, fixed minor bugs.
Fixed arguments for snappaths
Wrote tests for the dmd algorithm, and in doing so fixed many bugs in the
implementation of the dmd class.  All tests now pass, and each test was
verified against the same computation in Matlab (the test data were saved to
file and imported in Matlab for verification).

Other minor bug fixes/improvements were also made along the way:
-Some inconsistency in the use of numpy arrays vs matrices was fixed in the
bpod class.
-The util.load_mat_text method was updated to allow for complex valued
matrices.
-The methods for computing multiple inner products (modaldecomp class) were
updated so that a single string, rather than a list of strings, is also
allowable.
-Optional arguments for a list of snapshot paths were added to the
compute_decomp and compute_modes functions in the dmd and pod classes.  This
had previously only been changed in the bpod class.