Added mpi barriers so that the testfiles directory is cleaned after testing.
This had been a problem before because processors would delete data that
other processors still needed. This is now solved with a proper MPI.sync()
function. It is NOT tested itself, but does seem to produce the desired
results in testbpod.py and testmodaldecomp.py. I couldn't think of a simple
comprehensive test for this function to put in testutils.py.