Added test for case where snapPathList is a string.
A user may wish to take the inner product of one snapshot with many others,
e.g. in projection onto POD modes.  (Something like this is done in the
implementation of the DMD algorithm.)  This should be possible using the
methods already implemented in ModalDecomp, but required the ability to pass
in a single string rather than a list.  The ModalDecomp class checks for
this internally and changes the string to a list.

The update changed the tests for this class to test this functionality.