Added optional print statements for inner products (hankel matrix)
modaldecomp now has a data member verbose which can be true or false.
This can be set by the constructor or when calling _compute_inner_product_chunk().
BPOD also has this data member, again settable in the constructor, and passes
it on to _computer_inner_product_chunk() from _form_hankel().
This should be added to DMD and POD, but I'll leave that to others to do.
This is entirely optional, all previous code works and verbose will default to
False.