Removed extraneous constructor argument.
No reason to load ritzVals at construction.  The reason to do this for
buildCoeff is that we may only want to build the modes, but not compute a
decomposition.  In that case, we may load the buildCoeff from a file.
However, ritzVals does not get used except during compute_decomp, and at
that time it is always computed anyway.