Verbose now prints to stderr.
The verbose option now prints to stderr instead of stdout.  This is
preferred (by me, for now at least) because it is a frequent update
message.  stderr is unbuffered (at least in Python 2.7), so these messages
don't have to be flushed.  Also, the user cannot control the way these
statements print, so printing them to stderr allos the user to still print
messages to stdout without the verbosity argument interfering.