Fixed memory bug in dmd.
Now maxSnapsInMem is correctly pass to the internal pod object.  This was not
done previously, leading to memory issues.  Also, the tests for dmd have been
updated to work.  There was previously an error due to a constructor argument
that had been removed.  The tests now pass.  A test should be written to check
that all the arguments are correctly passed to dmd's internal pod object.