Basically a new version of modaldecomp, not backwards compatible.
This push contains many changes to mostly to bpod and modaldecomp:

maxSnapsInMem has been replaced by numNodes and maxSnapsPerNode.
This grouping makes more sense because maxSnapsPerNode typically doesn't
change, but numNodes does and can easily be passed as a command line argument from a submission script for the parallel case.

modes and snapshots are now both types of "fields", variable names have
been changed to account for this, such as load_field and save_field rather than load_snap and save_mode.

The default number of fields in memory at once is the minimum, 2, so that nodes aren't crashed. If the default isn't changed a warning is printed since only loading 2 at a time can be very inefficient.

MPI.find_proc_assigments now does not give an error if there are unused procs, just prints a warning. It is now up to the developer of a class to deal with the case of unused processors.

modaldecomp now uses _lin_combine_fields to form linear combinations of fields. In particular this is used to compute modes, since they are just a linear combination of snapshots. This method is also parallelized in a more efficient way than the previous _compute_modes was. Each processor is now only responsible for a set of snapshots (input fields), unlike before where each processor would read every snapshot (input field).

The tests have been updated for BPOD and ModalDecomp classes.

DMD and POD have not been touched, and need to be updated to work with these new changes. BPOD is working and can be looked at as an example.

The beginnings of an evaluate_and_bcast function are in util.MPI, but this isn't working yet.