Added compute_inner_product_matrix method to ModalDecomp base class.
We may often want to compute the inner product between two different sets of
snapshots.  For instance this is done to compute the Hankel matrix, and at one
stage of the DMD algorithm.  This would also be useful for projecting a set
of snapshots onto some modes.  It makes sense to parallelize this general task
so that the user doesn't have to.

The BPOD class now calls this method instead of the old internal _compute_hankel
method.  Similarly, DMD now makes use of this method as well, and POD was
updated to work with this changed too.

Tests were written for this new functionality, and all tests (for everything,
not just modaldecomp) pass in both serial and parallel.  Due to the large number
of cases tested, the modaldecomp tests take a while.  On Yellowstone they took
about 2 min in serial and 4 min in parallel.  The reason they are slower in
parallel is because the number of fields that can be loaded on a node is not
changed when the number of processors is increased, so it behaves as if we are
operating with less available memory.

A lot of minor formatting changes were also made, and a new warning is printed
if the number of fields that can be loaded per node is too small for the number
of procs (need at least 2 per proc).

Finally, the directories where test data are stored have been renamed to avoid
competing with source files when using autocompletion.