Added method for computing symmetric inner product matrices (serial).
When an inner product matrix is symmetric, only the upper triangular entries
need to be computed.  The rest can be filled in.  This was implemented in serial
in _compute_symmetric_inner_product_chunk.  The inner product matrix test was
extended to test this new functionality, and all tests pass in both serial and
parallel.