Rewrote method for computing symmetric inner product matrix and updated method for assigning processor tasks.
For memory/efficiency purposes, it was most useful to have a method that
computes an upper triangular block (not necessarily square) of inner products.
This was added and the method for computing a symmetric square block was
removed.

Processor tasks can now be given weights, which will affect how tasks are
assigned.

All of this new functionality has been tested in serial and parallel.

Old tests still pass too.