Major changes, no modaldecomp base class, added fieldoperations.
FieldOperations now has all of the functionality that ModalDecomp did,
and it exists as a seperate class so it can be used to compute the mean
flow or in ERA, for example. The interface is mostly unchanged. Some
optional arguments have been removed from some functions for clarity.

There is now a default MPI class instance, util.MPIInstance, that should
be used everywhere instead of making a new MPI class. This way, MPI is a
"singleton" class, there is only one instance, and it's almost
like a global variable.