NEW BRANCH - shared memory, not fully parallelized but somewhat functional
This new branch is only parallelized in a shared memory sense. All distributed memory
has been disabled for this case. In theory, shared memory should be very efficient
within a node. Only fieldoperations and bpod are currently usable and tested, ie POD
and DMD are not functional as far as I know (haven't tried them). The util.py has
been updated substantially and some of it should eventually be moved to even the
distributed version. Eventually this will be a hybrid parallelization (dist. and sh mem),
but mpi4py and multiprocessing modules are not interacting well right now and it will
take substantial work to make them work together, if it's possible at all.