Fully working shared memory, except inner products
The new version contains fully functional shared memory, which
mostly speeds up saving and loading. Currently inner products are
not parallelized in sh mem because it makes them slower, but I emailed
the numpy list and hopefully we will fix this in the future.
For now, this method should be significantly faster and less
wasteful than the distributed case.

Next steps are to make this work with both distributed and shared
memory and fix the inner products.