Moved creation of multiprocessing.Pool instance to within the fieldoperation class.
This solved two strange bugs Jon noticed.
The first was that there were multiprocessing-related errors when using a load_field
function that was defined in the same file that had the __main__.
The second was that there were multiprocessing-related errors when there were two
instances of FieldOperations and the pool.map function was used.