Made snapPaths list more flexible by not requiring a list of strings.
Now snapPaths can be a list of any object that is usable by functions such as
load_fields.  Previously, we checked if the object was a string or not, and if
it was a string, converted it to a list containing a single string.  Now we
check if the object is a list, and if not, convert it to a list containing said
object.

There is a possibility that functions such as load_field take lists as
arguments, and so snapPaths should be a list of lists.  In this case, our check
will not work, but for now we are not worrying about this case.

A new file called runalltests was added, to aid in development by running all
tests with one command.  Right now bpodltirom tests are not included, as they
don't seem to all pass yet.