Added options to load, save, compute inner products in either parallel or serial.
Tests show that for some operations, running in serial is faster.  For
instance, when the inner product is written using Python functions only, in a
brute force method (mutpliplying and adding elements in for loops), we get a
speed up approximately equal to the number of processors.  However, when a numpy
function is used to define the inner product, parallelization actually causes
the code to run slower.  The cause of this is as of yet unknown.  We think
perhaps it may have to do with some sort of internal parallelization within
numpy.

As far as saving and loading goes, it was observed in some cases that loading in
serial was faster than loading on many cores.  However, sometimes loading on
some small number of cores was faster than a single core, though scaling it up
eventually caused a slow down.  We suspect that here the problem is caused by
the fact that our data is generally stored all on one physical drive.  Trying to
load multiple datasets at once causes the drive head to have to skip around,
which may actually make the entire process slower.

In light of this, we have set options for each function call in which the user
can choose which operations should be done using shared memory.  By default,
everything is set to run in parallel, though we may wish to change this in the
future.

All tests have been updated and pass.  The tests are slower now, because every
combination of serial/parallel operations is tested, leading to 4x the number of
tests.