Parallelized symmetric inner product matrix computation.
This is implemented and passes tests.  As expected, only ~half as many inner
products are computed.  However, the choice of parallelization method does not
scale well with respect to the number of loads that must be done.  As such, this
should be re-parallelized.  Right now the POD method is implemented by simply
calling the compute_inner_product_mat.

The repository was also cleaned up. Files in doc/_build are now ignored, as they
may not be updated at every revision.  The user can simply build on his/her own.
In the user guide directory, the extraneous Latex files (*.aux, etc.) are now
ignored.  Only the Latex source is tracked.  The pdf is not because it may not
necessarily be up to date.  The figures are all tracked though.