Implemented new method for symmetric inner product matrices.
This method uses a new parallelization that should be more memory efficient, and
should scale better. Benchmarking is currently being done. It is unclear if the
desired speed-up is achieved. However, the method is tested and correctly
computes the desired matrices. Due to the unknown scaling behavior, POD is still
implemented with the standard compute_inner_product_mat function.

Also, a previously undetected race condition was fixed, in which the non-
blocking sends are now paired with Wait() commands.  The benchmark.py script was
updated too, now with the functionality to test the symmetric_inner_product_mat
command, and instructions for how to run the profiler.