Forced parallel functions to exit simultaneously.
Previously, for lin_combine, if a processor was not responsible for saving any
fields, it could exite the lin_combine function before the other processors.
This could cause a race condition wherein proc 0 might be responsible for
saving the mean flow, for later use in POD.  Since proc 1 has nothing to save,
it will exit lin_combine and continue executing the program, which may refer to
the mean flow, which has not yet been saved.  Now all parallel operations in the
fieldoperations module are forced to sync before exiting, preventing this
particular type of race condition.