Added functionality for complex inner product matrices.
Previously, the inner product matrices were allocated using the float datatype.
If inner products were complex, the imaginary part would be discarded.  Though a
warning was printed, this could eventually lead to incorrect results, for
instance in performing POD or DMD with complex datatypes.

It is assumed that an inner product function can only return a single datatype
(true mathematically, where the reals are considered to be a distinct field from
the complex numbers, not a subset).  A single inner product is taken to
determine the correct datatype, and this type is used when allocating the inner
product matrix.

A test was written to ensure that the correct datatype is selected.  However,
none was written to ensure the correct complex values are stored.  It is assumed
at this point that if the datatype is consistent with the output of the inner
product function, then there is no difference in testing any particular inner
product function compared with any other.  However, this could be changed in the
future for thoroughness.  (Likely through an additional nested loop in the test
for computing inner product matrices.)