Massive changes to naming and formatting, interface change!
Capital-letter genocide is complete. Now we use var_nam instead of varName
convention for variables and functions. Also, some formatting has been
improved so line breaks don't occur after '.'s and more is
implied line continuation. However, this is still an area that
needs *drastic* improvement. Having many indent levels, long variable
names, and an 80-column limit just cannot be done, one of these
needs to be sacrificed or the code will continue to be difficult
to read. This is especially bad in fieldoperations.py.