Major revision to ERA, util, and some docs (which were broken).
ERA:
Interface has changed slightly.
Heavy revisions to be entirely discrete-time oriented and simpler.
Currently it *only* works for data sampled at times:
[0 1 1 2 2 3 3 4 ...] or [0 1 2 3 4 ...], NOT [0 1 P P+1 2P 2P+1 ...].
The general P-sampled data results in models which aren't accurately balanced.
This might be OK - we're not sure, but it is best to avoid this format for now.
The load_impulse_outputs function is now in util, but may be moved again to a more
appropriate place.

util:
Added a few functions to help testing ERA, such as drss, impulse, and solve_Lyapunov.

docs:
sphinx-build doc doc/_build command wasn't working, so redid the docs. This
needed to be done anyway since we were still using some inherited pycontrol
files.