More interface changes to ERA.
Now less restrictive about the time_steps.
It is not necessary to give the ERA class any information about the time steps.
Some functions are now moved outside of the ERA class and are available mostly for
convenience, they are not essential.
These are for converting time step arrays and converting equally spaced data
to sampled data.