Updated POD.  Found memory problem.
In POD, the non-symmetric inner product routine was being called mistakenly.
This line was commented out so that the faster, symmetric routine is now called.

Also, some BPOD computations were crashing due to memory errors.  This occured
for runs where there were more than 1 chunk in the inner product matrix.  As
such, it may be necessary to clear the memory (using del) at the end of each
loop iteration.  A note about this was added to the todo file.