Changed save/load_mat to put/get_mat, and args to source/dest.
Same as done with get_field and put_field, now put/get_mat
are more general and one need not write to file to use
modred at all. The documentation and tests could still
be updated to better reflect these changes, but all tests
do pass.

Also working on writing a quickstart and intro in the
sphinx documentation.