compute_modes now takes a list of mode_dests, not one.
Before mode_dest was something like 'mode%d.txt', but this
doesn't make sense anymore because it need not be a filename.
Instead, mode_dests is a list of destinations to put modes, just
like we do for fields. This is more flexible and consistent.
All tests pass. Some documentation is updated as well.