Changes from before modred repo corruption.
The modred repo somehow became corrupted, and so
we had to clone from an earlier revision.
These are all the changes that had accumulated from
the previous revision. It is 1 revision, plus a few others
that we were preparing to commit, lumped into one.

The SimpleUse classes and tests have been added, as has
an example that uses them. There is no test for the example
script currently.

Small changes to the other tests dealing with adding paths
to sys.path for imports.

The user's guide has the figures for scaling removed because
they contain mistakes currently.