Full BPOD example. Tests of example scripts. Updated sphinx docs.
I added a full BPOD example with a non-uniform grid and save/loading.
It's the only example that uses modred the way we originally intended
with saving/loading and in parallel, but it may be sufficient.

I added tests of these scripts that simply run them. If there is an
error, then the tests show that. Since the scripts do not return anything,
I didn't know what else to test. This at least catches if there has been
an interface change that did not get fixed in the script.

I also updated the sphinx intro docs with the first few sections of the
user's guide.