Redid the parallelization of IP mat and lin combine.
Before we didn't make full use of find_assignments, instead
using some rounding and division. I removed the parts of this
that were unnecessary, and now make better use of find_assignments.

Still there is some division because the loops have to have
the same number of iterations for each processor. Thus we can't
exactly do the usual "for task in tasks" type loop, since
there are different numbers of tasks for each processor in general.