Changes accounting for Clancy's 1st comments.
The "compute_and_return" methods are now available.

No longer take put_vec/get_vec/inner_product as constructor
args, instead take a vecdefs class/module which has these
as an arg to VecOps constructor, and POD, BPOD, and DMD.

Many updates to docs, now a quickstart with mostly examples,
then another part with details for more advanced users.
This could still be cleaned up more.

Fix to tests so benchmark is tested.
Example scripts are tested properly -- before they tried to write
to the dir of the script which could be in the site-packages dir
which is write protected. Now they write to the CWD.

Added vecdefs module with some example classes and functions.
More cases might be helpful.

Updaated interface to BPODLTIROM to allow for return mats
or put_mat, just like in POD, BPOD, and DMD.