Fixed bug in symmetric inner product matrix routine.
In computing symmetric inner products, there are a few elements that are
computed twice, both in the upper triangular portion and the lower
triangular portion of the matrix.  To detect this, there had previously been
a check to see if any lower triangular element (i, j) was equal to the upper
triangular element (j, i).  If so, it was essentially marked as duplicate.

The problem occurs when the inner product function is not perfectly
symmetric, as then this test will not detect approximate duplicates.  This
was corrected by doing a better check, which when considering if (i, j) has
a duplicate, simply checks if (j, i) is nonzero.  If (i, j) is nonzero and
(j, i) is also nonzero, then we have a duplicate and only one of the
elements is used.

This check is equally valid and perhaps even more direct that the previous
equality check.  The only concern is that as of right now, there is not a
way to easily test this, as we generally assume perfectly symmetric inner
product functions.