Updated DMD to no longer use POD. Fixed vectors bug in __eq__.
DMD no longer uses POD to compute the decomposition, instead
following the procedure in the modred paper. The variables
need more descriptive names than what they have now.

THere was a bug when using base_vec_handles that were
InMemoryVecHandles relating to a NoneType (when no base_vec
given). This is now fixed by checking the type first in the
__eq__ method.