Updated variable names in DMD class.
Now names are again descriptive.  Combined smaller steps into one large step in
computing build_coeffs, which is a little hard to read, but otherwise fine.
This avoids the need for intermediate variable names.  Another option would be
to update this variable step-by-step, such that its true value is not attained
until the last step.  However, we have not used this strategy anywhere else in
the library, so it was avoided here.