Fixed potential bug in DMD/POD.
Now eigh wrapper in util can deal with matrices that are expected to be positive
definite.  In this case, if negative eigenvalues are encountered, the tolerance
is increased such that only positive eigenvalues are returned.  This is the case
encountered in the POD and DMD algorithms, when decomposing the correlation
matrices.

Implementation could be improved.  Right now, the call to parallel's
call_and_bcast is done using a hardcoded tolerance value, but should use eigh's
default tolerance.