Generalized DMD to work with non-sequential datasets.
Now DMD can take a sequential dataset, in which the data are partitioned into
two overlapping datasets, or a non-sequential one.  This is in line with the
general framework to be described in Jonathan Tu's PhD thesis, and also a paper
to be submitted soon.  Tests were written and pass in serial and parallel.

A few minor changes were made to the lin_combine method in vectorspace, in which
undesired behavior was previously allowed.  Now the tests checking the size of
the basis vector set and the build coefficient matrix are more stringent.