Fixed bug in DMD.
The error check on the size of the build_coeff matrix in DMD was incorrect.  It
was checking the number of columns of the build_coeff matrix instead of the
number of rows.  (It is the number of rows that should match the number of
vec_handles.)

No test was written for this, because if the data are full rank, the build_coeff
matrix is square, and this error is not observed.  To create this error would
require generating data that is rank deficient, whereas the tests use full rank
data.

However, in practice, this bug will be caught by error checks in lin_combine, so
computations done prior to this bug fix are likely ok.  Furthermore, if the
computation proceeded, then the build_coeff matrix was square.  Otherwise, the
computation would have failed.