Major revisions to DMD class, other minor changes.
Revised DMDHandles class to match new DMD theory in Tu et al, JCD, 2014.  Now
the build coefficients are produced for both the exact and projected DMD modes.
Instead of computing mode norms, terminology has been changed so that spectral
coefficients are computed instead.  (The actual values are the same as the
previously computed mode norms though.)  Also, there is now a method for
projecting the data vectors onto the DMD modes.  On a lower level, more data are
now stored as internal attributes, which will allow for more error checking and
advanced computations.

Other changes including the now consistent use of "eigvals" and "eigvecs" in all
files.  As part of the change to DMD, there is a new wrapped function in util
that computes both left and right eigenvectors of a matrix, sorting them and
scaling them to yield a biorthogonal set.