Updated biorthogonal eigendecomposition.
Now method doesn't crash if the left and right eigenvalues don't match.
Instead, it prints a warning.  Both the left and right eigenvalues are returned,
so that the user can error check.  The DMD class was updated to reflect these
changes, and the user can go in and check the left eigenvalues if desired.