Updated tests for util.svd and util.eigh.
SVD is now tested by checking that the singular vectors and singular values
satisfy the appropriate eigenvector/eigenvalue relationships.  This can be
thought of as the definition of an SVD, and other SVD properties can be derived
from it.  Thus we don't need to explicitly check that the singular vector
matrices are unitary, which had been causing problems.  The old test has now
been removed completely.

Also, now the tolerances in testing util.svd and util.eigh are determined during
testing, to ensure that truncation always occurs (otherwise the test would not
be testing the tolerance features).