Changed order of outputs from compute_decomp in BPOD.
Decided to emphasize the Hankel singular values, which have an important
physical meaning, over the the left and right singular vectors of the Hankel
matrix, which are used mostly as a means to other computations (like computing
the direct and adjoint modes).  This new ordering also better matches the
order of returned values for the BPOD matrix method.