parallel is now a module only, and no longer a class.
Suggested usage was to make use of the default instance of Parallel anyway, as
if it were a singleton.  This change enforces that usage.  There was really no
need to have the ability to create multiple instances of Parallel objects
anyway.  Now, users should be careful, as variables defined within parallel are
essentially global, so they should only be modified with extreme care.