Correctly handle case of mode_indices=None.
In VecSpaceMatrices, the None value was not handled as in other VectorSpace
classes.  Now if None is passed in the coefficient matrix is unaltered, and if
indices are passed in, it is sliced.  Previously, it was sliced no matter what,
and slicing with a None value leads to unintended results.