Fixed BPOD matrix methods test.
This test was not working before and was skipped all the time.  It now works
and passes 100/100 times in both serial and parallel.  However, due to the
intracacies of multiplying numpy matrices, and the type of data generated by
BPOD, the tests only pass with this level of reliability if the test is
implemented the same way as BPOD is implemented in modred itself.  This goes as
far as the order in which matrix multiplication is done (groupings) and whether
or not intermediate variables are used.  It also applies to whether or not the
final matrices are sliced, or if factor matrices are sliced ahead of time so
that slicing the product is unnecessary.

This sensitivity is a big downside of how the tests are implemented, as they
really are comparing to mirrored implementations of the same method.  It would
probably be better to test mathematical properties of BPOD, rather than
comparing two different implementations.