Adjusted inner product weights, SVD tolerances.
By adjusting inner product weights, states are now more evenly accounted for.
Previously, the inner product heavily emphasized some states over others, which
likely lead to the presence of highly uncontrollable/unobservable states.  This
change in inner products leads to higher pass rates.

The relative SVD tolerance was tightened in computing the Hankel matrix.  This
filters out states that are highly uncontrollable/unobservable.  The BPOD
properties are then tested to a tight tolerance on the remaining states.  This
increases the pass rate significantly and seems to be a reasonable choice, as
in practice users would be truncating uncontrollable/unobservable states
anyway.  The choice of 1e-6 as a minimum relative magnitude between the
smallest and largest HSVs is deemed reasonable as this allows for a state that
is 1 million times less controllable/observable than another.  While the tests
don't measure whether or not our methods produces correct modes for
(relatively) smaller HSVs, users of BPOD should not necessarily expect BPOD
properties to hold for very small HSVs anyway.  So this seems to be a
reasonable tradeoff between making sure the passes enough to give users
confidence, but with reasonable tolerances that make the test meaningful.