Array methods for POD, BPOD, DMD, TLSqrDMD now return a dictionary of auxiliary data.
Using a dictionary for auxiliary data avoids problems with choosing a natural
ordering for the return values when the auxiliary data are returned or not.
This was especially problematic for DMD, where there were auxiliary matrices as
well as auxiliary modes.

Tests were updated to match the new interface and all pass.