Force input data to be arrays.
Went through all modules and called np.array() on input arguments to force data
to be arrays.  This will avoid bugs due to users passing in matrices.  Was
conservative about this, calling np.array() in all functions, even when those
functions called on subfunctions that were already forcing data to be arrays.
This way, we are not relying on the correct implementation of the subfunctions.